backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
langchain
langchain-google-genai
langchain-openai
langchain-community
chromadb
langchain-chroma
sentence-transformers
pypdf
tenacity
httpx
orjson
pytest
